feat: product decrease and increase

+++ src/actions/descrease-cart-product/index.ts
@@ -8,12 +8,12 @@ import { db } from "@/db";
 import { cartItemTable } from "@/db/schema";
 import { auth } from "@/lib/auth";
 
-import { removeProductFromCartSchema } from "./schema";
+import { decreaseCartProductSchema } from "./schema";
 
-export const removeProductFromCart = async (
-  data: z.infer<typeof removeProductFromCartSchema>,
+export const decreaseCartProduct = async (
+  data: z.infer<typeof decreaseCartProductSchema>,
 ) => {
-  removeProductFromCartSchema.parse(data);
+  decreaseCartProductSchema.parse(data);
 
   const session = await auth.api.getSession({
     headers: await headers(),
@@ -40,5 +40,15 @@ export const removeProductFromCart = async (
     throw new Error("Unauthorized");
   }
 
-  await db.delete(cartItemTable).where(eq(cartItemTable.id, cartItem.id));
+  if (cartItem.quantity === 1) {
+    await db.delete(cartItemTable).where(eq(cartItemTable.id, cartItem.id));
+    return;
+  }
+
+  await db
+    .update(cartItemTable)
+    .set({
+      quantity: cartItem.quantity - 1,
+    })
+    .where(eq(cartItemTable.id, cartItem.id));
 };
